[inspect] Move data to a domain model
