+++ scat/src/main/java/scat/domain/model/Region.java
@@ -1,4 +1,4 @@
-package scat.data;
+package scat.domain.model;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -12,27 +12,20 @@ import javax.validation.constraints.NotNull;
 @Data
 @Entity
 @EqualsAndHashCode(of = "id")
-public class School {
+public class Region {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Column(nullable = false)
     private String name;
 
-    @Column
-    private Integer number;
-
     @NotNull
     @ManyToOne(fetch = FetchType.EAGER, optional = false)
-    @JoinColumn(name = "type")
-    private SchoolType type;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.EAGER, optional = false)
-    @JoinColumn(name = "city")
-    private City city;
+//    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "country")
+    private Country country;
 
 }
