+++ scat/src/main/java/scat/domain/service/CountryService.java
@@ -1,6 +1,6 @@
 package scat.domain.service;
 
-import scat.data.Country;
+import scat.domain.model.Country;
 import scat.domain.service.dto.CountryInput;
 
 /**
