+++ scat/src/main/java/scat/web/support/json/CityJsonSerializer.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import org.springframework.boot.jackson.JsonComponent;
 import org.springframework.boot.jackson.JsonObjectSerializer;
-import scat.data.City;
+import scat.domain.model.City;
 
 import java.io.IOException;
