+++ scat/src/test/java/scat/batch/address/AddressWriterTests.java
@@ -10,9 +10,9 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
 import scat.Entities;
 import scat.TestConfig;
-import scat.data.City;
-import scat.data.Country;
-import scat.data.Region;
+import scat.domain.model.City;
+import scat.domain.model.Country;
+import scat.domain.model.Region;
 import scat.repo.CityRepository;
 import scat.repo.CountryRepository;
 import scat.repo.RegionRepository;
