+++ scat/src/test/java/scat/batch/school/SchoolWriterTests.java
@@ -10,9 +10,9 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
 import scat.Entities;
 import scat.TestConfig;
-import scat.data.City;
-import scat.data.School;
-import scat.data.SchoolType;
+import scat.domain.model.City;
+import scat.domain.model.School;
+import scat.domain.model.SchoolType;
 import scat.repo.CityRepository;
 import scat.repo.SchoolRepository;
 import scat.repo.SchoolTypeRepository;
