+++ scat/src/test/java/scat/web/RegionControllerTests.java
@@ -12,8 +12,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.RequestBuilder;
 import scat.Entities;
 import scat.TestConfig;
-import scat.data.Country;
-import scat.data.Region;
+import scat.domain.model.Country;
+import scat.domain.model.Region;
 
 import static java.lang.String.format;
 import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8;
