+++ scat/src/test/java/scat/web/search/CountrySearchTests.java
@@ -10,7 +10,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.transaction.annotation.Transactional;
 import scat.Entities;
 import scat.TestConfig;
-import scat.data.Country;
+import scat.domain.model.Country;
 import scat.repo.CountryRepository;
 
 import java.util.List;
